requirements.txt
pycairo
numpy
iio
scipy
scikit-image